Minimal Installation of the Headscale Server
This patch consists of a minimal installation of the headscale server
with some minor configuration changes to make the config more or less
reaonable as a starting point.

I also enabled Debians auto updates, although I am not quite sure, if
they are periodically called, so we have to keep an eye on this for a
day or week.

Headscale server seems to start correctly, but I did no real testing
there, because I did not want to temper with my router for a test
server.